plan-my-outings/backend/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-mail==0.9.1
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0
werkzeug==2.3.7
pyjwt==2.8.0
tabulate==0.9.0
